requirements.txt
streamlit
python-dotenv
boto3
requests
streamlit-lottie
streamlit-antd-components
streamlit-option-menu
streamlit-modal
streamlit-timeline
streamlit-calendar
streamlit-extras
